requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
wikipedia
pyjokes
requests
langchain-community